requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyMuPDF==1.22.5
requests
streamlit
sentence-transformers
scikit-learn
numpy